Refactored project structure
Added SportStats Trait
