+++ src/main/scala/stats/SportStats.scala
@@ -1,0 +1,10 @@
+package stats
+
+import org.apache.spark.sql.DataFrame
+
+trait SportStats {
+  var dataFrame: DataFrame
+  def getGeneralInfo: DataFrame
+  def getIndividualStats: DataFrame
+  def getEventStats: DataFrame
+}
