+++ src/main/scala/stats/TeamSportStats.scala
@@ -1,0 +1,7 @@
+package stats
+
+import org.apache.spark.sql.DataFrame
+
+trait TeamSportStats extends SportStats{
+  def getTeamStats: DataFrame
+}
